// apps/admin-web/src/.generated/api/model/authControllerRefreshV1200.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Admin API
 * OpenAPI spec version: 1.0.0
 */
import type { AuthControllerRefreshV1200Meta } from './authControllerRefreshV1200Meta';
import type { RefreshResponse } from './refreshResponse';

export type AuthControllerRefreshV1200 = {
  success: boolean;
  statusCode: number;
  path: string;
  requestId: string;
  timestamp: string;
  data: RefreshResponse;
  message?: string;
  meta?: AuthControllerRefreshV1200Meta;
};